Change host option to URL parsing instead of string
Code that handles this as a URL was added in ba678c5 but I neglected to
commit this line from a dirty work tree, leaving the state internally
inconsistent in the repository so commits in interim won't build the
cmd package successfully.

And this also adds the net/http import needed by 85ab384.

Doh.

Note to self: stop committing with a dirty tree. It's also about time to
add some tests for the cmd package so CI builds would fail.