Make the client constructor API nicer
In the case of a client for an API that isn't a public service with a
canonical URL, users are much more likely to set the host URL than a
custom HTTP client. So, make that the easier case.

Inspired by the go/build Gerrit client.

As a bonus, this is more robust about whether the user includes a
trailing slash or not.