Fix golint installation, version tools
golint changed its import path [1], and that along with the advent of
modules caused fallout [2, 3] that broke the `go get -u` installation in
our makefile/CI build.

The tools.go idiom is the currently favored approach for versioning
development tools with the module system [4, 5], in a way that `go mod
tidy` won't churn them from `go.mod` and the `+build` constraint keeps
them out of actual build products.

The tools still need to be `go install`ed, within a module `go get -u`
is not the thing to do anymore because it upgrades transitive deps of a
tool which may change the module's build. It takes like hours of reading
discussions to triangulate on these moving targets... [5, 6, 7, 8]

jfc how much of life have I spent following the fashion evolution of Go
dependency management

[1]: https://github.com/golang/lint/commit/c363707d68842c977f911634e06201907b60ce58
[2]: https://github.com/golang/go/issues/30455
[3]: https://github.com/golang/go/issues/30831
[4]: https://github.com/golang/go/wiki/Modules#how-can-i-track-tool-dependencies-for-a-module
[5]: https://github.com/golang/go/issues/25922
[6]: https://github.com/golang/go/issues/27653
[7]: https://github.com/golang/go/issues/27643
[8]: https://github.com/golang/go/issues/30515